requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-environ==0.7.0
django-heroku==0.3.1
django-markdownx==3.0.1
django-widget-tweaks==1.4.8
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
